Add Code of Conduct (#28)
* Add link to COC in `CONTRIBUTING.md`

We are about to add a Code of Conduct and it makes sense to surface that in the `CONTRIBUTING.md`.

* Add CODE_OF_CONDUCT.md

In the past Facebook didn't promote including a Code of Conduct when creating new projects, and many projects skipped this important document. Let's fix it. :)

**why make this change?:**
Facebook Open Source provides a Code of Conduct statement for all
projects to follow, to promote a welcoming and safe open source community.

Exposing the COC via a separate markdown file is a standard being
promoted by Github via the Community Profile in order to meet their Open
Source Guide's recommended community standards.

As you can see, adding this file will improve [the shortcuts-for-framer community profile](https://github.com/facebook/shortcuts-for-framer/community)
checklist and increase the visibility of our COC.

**test plan:**
Viewing it on my branch -
(Flarnie will insert screenshots)

**issue:**
internal task t23481323